requirements.txt
asgiref==3.3.1
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap4==2.3.1
djangorestframework==3.12.2
idna==2.10
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
unicode==2.8
unicodedata2==13.0.0.post2
urllib3==1.26.3
